move calls to addParty() and addEnemies() into init()
Also, change addParty and addEnemies to not themselves set this.party
or this.enemies, so they can be used more than once in a single combat,
if wanted/needed.

For example, maybe a boss summons new monsters, or a character jobs the
fight after a certain point in the combat, etc.